Health Bar
I have written a script for Health Bar. The health bar shows decrement when the player takes hits and current health decreases. The health bar is green slider and when the health is below 33% or the health of the player is critical the bar turns to red. I don't have a good coding background so the script might have  issues. If the code I wrote causes issues , we can remove it all.