Updated Game manager. (Pause menu-> settings menu.), Updated pauser menu UI.
Main menu to settings menu was taking to updated settings menu, but the resume menu to settings menu was taking to old settings menu. Few tweaks in the controls menu UI scrolling/scrollbar and scroll panel. With pause menu UI, placed the sprite with the book and changed the location of the text and the button.